Tune Feedback: Fixed bug in refresh delta tunes test
We are testing that the refresh_delta_tunes method raises
an exception when a tune is in an minor alarm. If the PV isn't
in an alarm the test still passes as the method rasies an
exception as the value of the PV is nan, and the timestamp is
to old! This is fixed in this commit.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/ioc/CS/CS-DI-IOC-09@131593 e099a375-04f9-0310-9d5f-a741eaff62e1